feat(aod_rx): integrate all components into main application
The main.c file has intentionally been reverted to a version that is
minimally different from the initial BLE AoD connectionless receiver
sample. "git diff" and "git difftool" will reveal minimally invasive
additions to the initial sample code. Tabular indentation has also been
preserved to avoid artifical differences. This aims to make it easy to
contrast this version against the initial sample code. Maybe this will
be appreciated by the next person working on this code. Unfortunately,
this version is not able to properly cycle between beacons. It tends to
sync to the same beacon indefinitely. It will repeatedly calculate
direction cosines, azimuth, and elevation from the first beacon it
encounters, but it will not be able to calculate a locator position
without direction cosines from a second beacon. The locator can be
tricked into syncing with a second beacon by power cycling both beacons.
Each beacon transition will result in a new locator position. There has
been a version of main.c capable of cycling between beacons, but that
version was rushed in preparation of a simple proof-of-concept
demonstration. The rushed modifications were reverted beacuse they were
not appropriate for the longterm development of the locator.

modified:   locator/src/main.c